Add Book components and coded it
